move single node defaults configs into its module
