+++ src/common/src/lib.rs
@@ -74,14 +74,12 @@ pub mod opts;
 pub mod range;
 pub mod row;
 pub mod session_config;
-pub mod single_process_config;
 pub mod system_param;
 pub mod telemetry;
 pub mod test_utils;
 pub mod transaction;
 pub mod types;
 pub mod vnode_mapping;
-
 pub mod test_prelude {
     pub use super::array::{DataChunkTestExt, StreamChunkTestExt};
     pub use super::catalog::test_utils::ColumnDescTestExt;
